Tests/LibWeb/Text/expected/HTML/img-src-set-twice-to-same-url.txt

naturalWidth: 2
naturalHeight: 1
complete: true
